+++ packages/react/src/components/radio-card-group/radio-card.tsx
@@ -14,7 +14,7 @@ import { v4 as uuid } from '../../utils/uuid';
 import { useDeviceContext } from '../device-context-provider/device-context-provider';
 import * as S from './styled-components';
 
-interface RadioCardProps {
+export interface RadioCardProps {
     checked?: boolean;
     children?: ReactNode;
     className?: string;
@@ -46,7 +46,8 @@ export const RadioCard : VoidFunctionComponent<RadioCardProps> = ({
     const { isMobile } = useDeviceContext();
     const id = useMemo(() => providedId || uuid(), [providedId]);
     const inputRef = useRef<HTMLInputElement>(null);
-    const containerRef = useRef<HTMLLabelElement>(null);
+    const labelRef = useRef<HTMLLabelElement>(null);
+    const containerRef = useRef<HTMLDivElement>(null);
     const [isInputChecked, setInputCheck] = useState(defaultChecked);
     const dataAttributes = useDataAttributes(otherProps);
 
@@ -74,45 +75,45 @@ export const RadioCard : VoidFunctionComponent<RadioCardProps> = ({
     }
 
     return (
-        <S.Label
+        <S.Card
             className={className}
             data-testid={`radio-card-${value}-container`}
             disabled={disabled}
             isMobile={isMobile}
             isChecked={isInputChecked}
             ref={containerRef}
-            onMouseDown={(e) => e.preventDefault()}
+            onMouseDown={(e) => {
+                labelRef.current?.click();
+                e.preventDefault();
+            }}
         >
-            <S.InputContainer
+            <S.HiddenInput
+                aria-describedby={`description-${id}`}
+                checked={checked}
+                data-testid={`radio-card-${value}-input`}
+                defaultChecked={defaultChecked}
                 disabled={disabled}
+                id={id}
                 isMobile={isMobile}
-                isChecked={isInputChecked}
-                key={`${name}-${value}`}
-            >
-                {label}
-                <S.HiddenInput
-                    aria-describedby={`description-${id}`}
-                    checked={checked}
-                    data-testid={`radio-card-${value}-input`}
-                    defaultChecked={defaultChecked}
-                    disabled={disabled}
-                    id={id}
-                    isMobile={isMobile}
-                    name={name}
-                    ref={inputRef}
-                    required={required}
-                    type="radio"
-                    value={value}
-                    onBlur={handleBlur}
-                    onChange={handleChange}
-                    onMouseDown={(e) => e.preventDefault()}
-                    {...dataAttributes /* eslint-disable-line react/jsx-props-no-spreading */}
-                />
-                <S.RadioInput disabled={disabled} isMobile={isMobile} />
-            </S.InputContainer>
-            <S.Description aria-hidden disabled={disabled} id={`description-${id}`} isMobile={isMobile} isChecked={isInputChecked}>
-                {children}
-            </S.Description>
-        </S.Label>
+                name={name}
+                ref={inputRef}
+                required={required}
+                type="radio"
+                value={value}
+                onBlur={handleBlur}
+                onChange={handleChange}
+                onMouseDown={(e) => e.preventDefault()}
+                {...dataAttributes /* eslint-disable-line react/jsx-props-no-spreading */}
+            />
+            <S.Label ref={labelRef} htmlFor={id}>
+                <S.Title isChecked={checked} disabled={disabled} isMobile={isMobile}>
+                    <S.RadioInput isChecked={checked} disabled={disabled} isMobile={isMobile} />
+                    {label}
+                </S.Title>
+                <S.Description aria-hidden disabled={disabled} id={`description-${id}`} isMobile={isMobile} isChecked={isInputChecked}>
+                    {children}
+                </S.Description>
+            </S.Label>
+        </S.Card>
     );
 };
